requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
arxiv
wikipedia
validators
youtube_transcript_api
pytube
pydantic
pydantic-core
unstructured
huggingface_hub
tiktoken
cassio
datasets
